requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
disnake==2.8.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
multidict==6.0.4
openai==0.27.2
Pillow==9.4.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.28.2
SQLAlchemy==2.0.7
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
